Move dataframe saving/loading code from Package to PackageStore
The logic belongs in PackageStore rather than Package because objects are owned by the store, and might not even have a package associated with them. (This will be the case once filtering is implemented.)
